+++ src/aui-video/js/aui-video.js
@@ -17,8 +17,6 @@ var Lang = A.Lang,
     TPL_VIDEO = '<video id="{id}" controls="controls" class="' + CSS_VIDEO_NODE + '" {height} {width}></video>',
     TPL_VIDEO_FALLBACK = '<div class="' + CSS_VIDEO_NODE + '"></div>';
 
-    WIN = A.getWin();
-
 /**
  * A base class for Video.
  *
@@ -52,19 +50,6 @@ var Video = A.Component.create({
      */
     ATTRS: {
 
-        /**
-         * The ratio used to resize the video for various screen sizes.
-         *
-         * @attribute aspectRatio
-         * @type {Number}
-         */
-        aspectRatio: {
-            valueFn: function() {
-                var instance = this;
-
-                return instance.get('height') / instance.get('width');
-            }
-        },
         /**
          * The required Flash version for the swf player
          *
@@ -250,7 +235,23 @@ var Video = A.Component.create({
             instance.publish('play');
             instance.publish('pause');
 
-            WIN.on(['orientationchange', 'resize'], instance._resizeVideo, instance);
+            instance._eventHandles = [
+                A.after(
+                    'windowresize',
+                    A.bind('afterWindowResize', instance)
+                )
+            ];
+        },
+
+        /**
+          * Destructor implementation.
+          * Lifecycle.
+          *
+          * @method destructor
+          * @protected
+          */
+        destructor: function() {
+            (new A.EventHandle(instance._eventHandles)).detach();
         },
 
         /**
@@ -268,9 +269,6 @@ var Video = A.Component.create({
                     roleNode: instance.get('contentBox')
                 });
             }
-
-            instance._removeBoundingBoxDimensions();
-            instance._resizeVideo();
         },
 
         /**
@@ -313,6 +311,25 @@ var Video = A.Component.create({
         },
 
         /**
+         * Remove the defined height and width from the bounding box.
+         *
+         * @method responsiveBoundingBox
+         * @publich
+         */
+        responsiveBoundingBox: function() {
+            var instance = this;
+
+            var boundingBox = instance.get('boundingBox');
+
+            boundingBox.setStyles(
+                {
+                    height: '',
+                    width: ''
+                }
+            );
+        },
+
+        /**
          * Create `source` element
          * using passed type attribute.
          *
@@ -329,25 +346,18 @@ var Video = A.Component.create({
         },
 
         /**
-         * Removes the inline styling for the boundingBox.
+         * Fired after the `windowresize` event.
          *
-         * @method _removeBoundingBoxDimensions
-         * @protected
+         * @method afterWindowResize
+         * @param {EventFacade} event
+         * @public
          */
-        _removeBoundingBoxDimensions: function() {
+        afterWindowResize: function(event) {
             var instance = this;
 
-            var bb = instance.get('boundingBox');
+            instance.responsiveBoundingBox();
 
-            var bbStyle = bb._node.style;
-
-            if (bbStyle.removeProperty) {
-                bbStyle.removeProperty('width');
-                bbStyle.removeProperty('height');
-            } else {
-                bbStyle.removeAttribute('width');
-                bbStyle.removeAttribute('height');
-            }
+            instance._setResponsiveDimensions(event);
         },
 
         /**
@@ -477,28 +487,43 @@ var Video = A.Component.create({
         },
 
         /**
-         * Resizes the video at various window sizes.
+         * Set the dimensions of the video player based on the window size.
          *
-         * @method _resizeVideo
+         * @method _setResponsiveDimensions
+         * @param {EventFacade} event
          * @protected
          */
-        _resizeVideo: function () {
+        _setResponsiveDimensions: function(event) {
             var instance = this;
 
-            var video = instance._video;
-            var videoInitialWidth = instance.get('width');
-            var videoInitialHeight = instance.get('height');
-            var videoParentWidth = video.ancestor('.video-content').width();
+            var height = instance.get('height');
+            var width = instance.get('width');
 
-            if (WIN.width() < videoInitialWidth) {
-                var newWidth = videoParentWidth;
+            var aspectRatio = height / width;
 
-                video.width(newWidth);
-                video.height(newWidth * instance.get('aspectRatio'));
-            } else {
-                video.width(videoInitialWidth);
-                video.height(videoInitialHeight);
+            var updatedHeight = height;
+            var updatedWidth = width;
+
+            var currentTarget = event.currentTarget;
+
+            var currentTargetHeight = currentTarget.height();
+
+            if (currentTargetHeight < height) {
+                updatedHeight = currentTargetHeight;
+                updatedWidth = currentTargetHeight * aspectRatio;
             }
+
+            var currentTargetWidth = currentTarget.width();
+
+            if (currentTargetWidth < width) {
+                updatedHeight = currentTargetWidth * aspectRatio;
+                updatedWidth = currentTargetWidth;
+            }
+
+            var video = instance._video;
+
+            video.width(updatedWidth);
+            video.height(updatedHeight);
         },
 
         /**