+++ src/aui-video/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-No registries yet.
+* [AUI-2112](https://issues.liferay.com/browse/AUI-2112) AUI video should be responsive
 
 ## [3.0.3](https://github.com/liferay/alloy-ui/releases/tag/3.0.3)